pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-direct-agent-v3"
version = "0.1.0"
description = "Autonomous Yandex.Direct management agent for 24bankrotsttvo.ru"
requires-python = ">=3.12"
authors = [{ name = "Mihail Maklov", email = "myaklov@yandex.ru" }]
dependencies = [
    "agents-core[observability,pg] @ git+https://github.com/mmnmaikl-sys/agents-core.git@b2fea6e8087c1a707754505d77767cd7358b67d4",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "psycopg[binary,pool]>=3.2",
    "python-telegram-bot>=21",
    "pydantic-settings>=2",
    "slowapi>=0.1",
    "Pillow>=10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3",
    "pytest-cov>=4",
    "pytest-postgresql>=6",
    "httpx>=0.27",
    "ruff>=0.6",
    "pyright>=1.1.380",
    "pre-commit>=3.8",
    "pip-audit>=2.7",
    "shellcheck-py>=0.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_runtime*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "ASYNC", "T20"]
ignore = [
    "S101",  # allow `assert` in tests and smoke
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S603", "S607", "S106", "S107"]  # test subprocess/assert/fake-secrets allowed
"scripts/**" = ["T201"]  # CLI scripts print to stdout/stderr by design

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["agent_runtime", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportMissingImports = "warning"
